Moved all the *State classes out of their modules in separate files, and placed all the modules in their own dirs.
