+++ lib/core/network/index.js
@@ -13,7 +13,7 @@
 var spawn = require('child_process').spawn;
 var Q = require('q');
 
-var logger = require('./../utils/log4js-wrapper').getLogger("Network");
+var logger = require('../../utils/log4js-wrapper').getLogger("Network");
 
 
 /**
