+++ lib/core/plugins/index.js
@@ -14,54 +14,41 @@ var fork = require('child_process').fork;
 
 var Q = require('q');
 
-var fspromised = require('./../utils/fspromised');
-var logger = require('./../utils/log4js-wrapper').getLogger("Plugins");
-var helpers = require('./../utils/helpers');
-var constants = require('../shared/constants/plugin');
-var PluginsState = require('../shared/plugins-state');
-var JsonConfigsUtils = require('./../utils/json-configs-utils');
+var fspromised = require('../../utils/fspromised');
+var logger = require('../../utils/log4js-wrapper').getLogger("Plugins");
+var helpers = require('../../utils/helpers');
+var constants = require('../../shared/constants/plugin');
 
-var NoopError = require('./plugins/noop-error');
-var StatusError = require('./plugins/status-error');
-var TypeError = require('./plugins/type-error');
-var PeriodicalWatchdogRunningPluginsTracker = require('./plugins/watchdog-running-tracker');
+var JsonConfigsUtils = require('../../utils/json-configs-utils');
 
-var pluginsProcessRelPath = '../process/plugin';
+var NoopError = require('./noop-error');
+var StatusError = require('./status-error');
+var TypeError = require('./type-error');
+var PeriodicalWatchdogRunningPluginsTracker = require('./watchdog-running-tracker');
 
+var pluginsState = require('./_state');
 
-/**
- * Extends PluginsState with a running plugins tracker
- *
- * @constructor
- */
-function CorePluginsState() {
-    PluginsState.call(this);
-
-    /**
-     * The purpose of this class is to consistently keep track of processes
-     * started by current LR Core process.
-     *
-     * Since plugins.json statuses and pids might be edited outside, we need
-     * to keep a reliable list of processes we started by ourselves and
-     * (eventually) sync it with the file.
-     *
-     * Without that we wouldn't have been able to stop processes started
-     * here and switched to 'off' outside.
-     *
-     * Moreover, we need to determine sometimes whether we have started
-     * the plugin or not in order to reject/redo the requested operation.
-     *
-     * @type {RunningPluginsTracker}
-     */
-    this._runningPluginsTracker =  // previously was 300 * 1000
-        new PeriodicalWatchdogRunningPluginsTracker(pluginsWatchdogTick, 60 * 1000);
+var pluginsProcessRelPath = '../../process/plugin';
 
 
-}
-util.inherits(CorePluginsState, PluginsState);
-
-
-var pluginsState = new CorePluginsState();
+pluginsState.onPluginRestartIsNeeded = function (pluginName) {
+    Plugins.kill(pluginName)
+        .catch(function (err) {
+            if (err instanceof NoopError) {
+                // already stopped plugin is not a failure
+                return;
+            }
+            throw err;
+        })
+        .then(function () {
+            // todo ?? set pid to "" in the global conf
+            return pluginsState.readPluginConfig(pluginName);
+        })
+        .then(function (pluginConf) {
+            return Plugins.startAsynchronousPlugin(pluginName, pluginConf);
+        })
+        .done();
+};
 
 
 /**
@@ -449,38 +436,6 @@ Plugins.restartAllActivePlugins = function () {  // todo was marked as DEPRECATE
         });
 };
 
-
-function pluginsWatchdogTick() {
-    pluginsState._runningPluginsTracker.each(function (pluginName, runningPlugin) {
-        if (runningPlugin.isRunning()) {
-            return;
-        }
-
-        if (runningPlugin.isSync) {  // we should watch for async plugins only
-            return;
-        }
-
-        logger.warn('[PLUGIN] - PluginChecker - ' + pluginName
-            + ' - No such process found - Restarting...');
-
-        Plugins.kill(pluginName)
-            .catch(function (err) {
-                if (err instanceof NoopError) {
-                    // already stopped plugin is not a failure
-                    return;
-                }
-                throw err;
-            })
-            .then(function () {
-                // todo ?? set pid to "" in the global conf
-                return pluginsState.readPluginConfig(pluginName);
-            })
-            .then(function (pluginConf) {
-                return Plugins.startAsynchronousPlugin(pluginName, pluginConf);
-            })
-            .done();
-    });
-}
 
 /**
  * Start async plugins which should be autostarted
