+++ lib/core/socat/index.js
@@ -13,55 +13,17 @@ var Q = require('q');
 var nconf = require('nconf');
 var fork = require('child_process').fork;
 
-var ConfigError = require('./../utils/config-error');
-var logger = require('./../utils/log4js-wrapper').getLogger("Socat");
-var Reverse = require('./reverse');
-var WAMP = require('./wamp');
-var Board = require('./board');
-var constants = require('../shared/constants/socat');
+var ConfigError = require('../../utils/config-error');
+var logger = require('../../utils/log4js-wrapper').getLogger("Socat");
+var Reverse = require('../reverse');
+var WAMP = require('../wamp');
+var Board = require('../board');
+var constants = require('../../shared/constants/socat');
 
-var socatProcessRelPath = '../process/socat';
+var socatProcessRelPath = '../../process/socat';
 
+var socatState = require('./_state');
 
-var socatState = new SocatState();
-
-/**
- * Socat subsystem state
- * Don't ever cache it's results, because they might be changed at any time.
- *
- * @constructor
- */
-function SocatState() {
-    this._socatPort = null;
-}
-
-
-/**
- * Returns false if SocatState is not loaded yet
- *
- * @returns {boolean}
- */
-SocatState.prototype.isLoaded = function () {
-    return this._socatPort !== null;
-};
-
-
-/**
- * Reloads Socat subsystem state
- *
- * @throws {ConfigError}
- * @param nconf_ might be omitted
- */
-SocatState.prototype.reload = function (nconf_) {
-    nconf_ = nconf_ || nconf;
-
-    var socatPort = nconf_.get('config:socat:client:port');
-    if (!socatPort) {
-        throw new ConfigError("config:socat:client:port can't be empty");
-    }
-
-    this._socatPort = socatPort;
-};
 
 /**
  * Socat subsystem singleton
