+++ lib/core/wamp/_state.js
@@ -14,14 +14,15 @@ var nconf = require('nconf');
 var urlparse = require('url').parse;
 var Q = require('q');
 
-var ConfigError = require('./../utils/config-error');
-var logger = require('./../utils/log4js-wrapper').getLogger("WAMP");
+var ConfigError = require('../../utils/config-error');
+var logger = require('../../utils/log4js-wrapper').getLogger("WAMP");
 
-var AutobahnConntestTcpkillWAMPConnection = require('./wamp/connection/autobahn-conntest-tcpkill');
-var Manage = require('./manage/index');
-var Board = require('./board');
+var AutobahnConntestTcpkillWAMPConnection = require('../wamp/connection/autobahn-conntest-tcpkill');
+var Manage = require('../manage');
+var Board = require('../board');
 
-var wampState = new WAMPState();
+
+var wampState = module.exports = new WAMPState();
 
 /**
  * WAMP subsystem state.
@@ -223,18 +224,3 @@ function _wampStateConnectionOnOpen() {
     //driversManager.restartDrivers();
 
 }
-
-
-/**
- * WAMP subsystem singleton
- */
-var WAMP = module.exports = {};
-
-/**
- * Returns WAMPState
- *
- * @returns {WAMPState}
- */
-WAMP.getState = function () {
-    return wampState;
-};
