+++ lib/core/wamp/index.js
@@ -10,22 +10,19 @@
 "use strict";
 
 
-var BoardState = require('../shared/board-state');
-
-
-var boardState = new BoardState();
+var wampState = require('./_state');
 
 
 /**
- * Board subsystem singleton
+ * WAMP subsystem singleton
  */
-var Board = module.exports = {};
+var WAMP = module.exports = {};
 
 /**
- * Returns BoardState
+ * Returns WAMPState
  *
- * @returns {BoardState}
+ * @returns {WAMPState}
  */
-Board.getState = function () {
-    return boardState;
+WAMP.getState = function () {
+    return wampState;
 };
